glide: Change easing curve for closing animation
The previous update to aeroglide ended up with complaints from users
that the closing animation has weird timing, feeling like it is delayed.
This had to do with the fact that the easing curve being used was based
on a cubic curve, which has a slow initial start.

The closing animation now uses a quadratic easing curve which should
give reasonably good results without feeling like the animation is slow
or delayed.